Action classes created, but they are not finished.

+++ Engine/src/properties/property/api/PropertyType.java
@@ -1,5 +1,5 @@
 package properties.property.api;
 
 public enum PropertyType {
-    INT, FLOAT, BOOLEAN, STRING
+    INT, DOUBLE, BOOLEAN, STRING
 }
